+++ test/analyzer/sql/test_is_null_rule.py
@@ -16,7 +16,7 @@
 @desc:
 """
 import unittest
-from src.handler.analyzer.sql.rules import IsNullRule
+from src.handler.analyzer.sql.rules.review.is_null import IsNullRule
 from sqlgpt_parser.parser.oceanbase_parser import parser
 from src.handler.analyzer.sql.rules.level import Level
