+++ test/analyzer/sql/test_sellect_all_rule.py
@@ -17,7 +17,7 @@
 """
 import unittest
 from sqlgpt_parser.parser.oceanbase_parser import parser
-from src.handler.analyzer.sql.rules import SelectAllRule
+from src.handler.analyzer.sql.rules.review.select_all import SelectAllRule
 
 
 class TestSelectAllCase(unittest.TestCase):
