+++ test/common/test_config_helper.py
@@ -17,7 +17,7 @@
 """
 import unittest
 from unittest import mock
-from src.common import ConfigHelper
+from src.common.config_helper import ConfigHelper
 
 
 class TestConfigHelper(unittest.TestCase):
